+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Support for `instance` and `special` validators for Valibot, which now enables File validation for Valibot!
+- `taintedMessage` didn't always work when navigating with the History API.
 
 ## [2.9.0] - 2024-03-12